+++ src/main/java/com/grepp/spring/app/model/quiz/dto/internal/QuizDto.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.quiz.dto.maybeResponse;
+package com.grepp.spring.app.model.quiz.dto.internal;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
