+++ src/main/java/com/grepp/spring/app/model/quiz/dto/internal/QuizGenerationDto.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.quiz.dto.maybeResponse;
+package com.grepp.spring.app.model.quiz.dto.internal;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
