+++ src/main/java/com/grepp/spring/app/model/quiz/dto/internal/QuizProjection.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.quiz.dto.maybeResponse;
+package com.grepp.spring.app.model.quiz.dto.internal;
 
 import lombok.Getter;
 import java.util.List;
