+++ src/main/java/com/grepp/spring/app/model/quiz/dto/response/QuizListResponse.java
@@ -1,6 +1,6 @@
 package com.grepp.spring.app.model.quiz.dto.response;
 
-import com.grepp.spring.app.model.quiz.dto.maybeResponse.QuizDto;
+import com.grepp.spring.app.model.quiz.dto.internal.QuizDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import java.util.List;
