+++ src/main/java/com/grepp/spring/app/model/quiz/repository/quizSetRepository/QuizSetRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.grepp.spring.app.model.quiz.repository.quizSetRepository;
 
-import com.grepp.spring.app.model.quiz.dto.maybeResponse.QuizProjection;
+import com.grepp.spring.app.model.quiz.dto.internal.QuizProjection;
 import java.util.List;
 
 public interface QuizSetRepositoryCustom {
